+++ tests/testthat/test-model_parameters.htest.R
@@ -120,7 +120,7 @@ test_that("model_parameters-t-test standardized d", {
   )
 })
 
-test_that("model_parameters-t-test standardized d", {
+test_that("model_parameters-t-test standardized d-2", {
   mp <- model_parameters(
     t.test(mtcars$mpg ~ mtcars$vs),
     es_type = "cohens_d",
@@ -200,4 +200,71 @@ test_that("model_parameters-htests does not fail for fisher test > 2x3", {
   expect_named(out, c("Chi2", "df", "p", "Method", "Alternative"))
   expect_true(is.na(out$Chi2))
   expect_equal(out$p, 8e-05, tolerance = 1e-4)
+})
+
+test_that("model_parameters-htests no hard-coded formatting for proportions tests", {
+  smokers <- c(83, 90, 129, 70)
+  patients <- c(86, 93, 136, 82)
+  out <- model_parameters(prop.test(smokers, patients))
+
+  expect_identical(
+    capture.output(out),
+    c(
+      "4-sample test for equality of proportions without continuity correction",
+      "",
+      "Proportion                        | Chi2(3) |     p",
+      "---------------------------------------------------",
+      "96.51% / 96.77% / 94.85% / 85.37% |   12.60 | 0.006",
+      "",
+      "Alternative hypothesis: two.sided"
+    )
+  )
+  expect_equal(out$Estimate_1, 0.96512, tolerance = 1e-4)
+  expect_named(
+    out,
+    c(
+      "Estimate_1",
+      "Estimate_2",
+      "Estimate_3",
+      "Estimate_4",
+      "Chi2",
+      "df",
+      "p",
+      "Method",
+      "Alternative"
+    )
+  )
+
+  # for penguins data
+  skip_if(getRversion() < "4.5.0")
+  data(penguins, package = "datasets")
+  out <- model_parameters(prop.test(table(penguins$sex)))
+  expect_identical(
+    capture.output(out),
+    c(
+      "1-sample proportions test",
+      "",
+      "Proportion |       95% CI | Chi2(1) | Null_value |     p",
+      "--------------------------------------------------------",
+      "49.55%     | [0.44, 0.55] |    0.01 |       0.50 | 0.913",
+      "",
+      "Alternative hypothesis: true p is not equal to 0.5"
+    )
+  )
+  expect_equal(out$Estimate, 0.4955, tolerance = 1e-3)
+  expect_named(
+    out,
+    c(
+      "Estimate",
+      "CI",
+      "CI_low",
+      "CI_high",
+      "Chi2",
+      "df",
+      "Null_value",
+      "p",
+      "Method",
+      "Alternative"
+    )
+  )
 })